renames things to they make more sense
