+++ frontend/src/pages/Search/DashboardPage.tsx
@@ -1,1 +1,1 @@
-export { Dashboard as SearchPage } from './Dashboard';
+export { Dashboard as SearchPage } from './Inventory';
