+++ frontend/src/pages/Search/Inventory.tsx
